M0 spike + M1/WS-1: restore clean build, byte-exact frozen CPUID profile, zero-rule, close native-leak paths
- M0: remove fabricated WHP syscall bitmap (never existed in any SDK), fix
  missing targets in build files, missing includes, export-ID mismatches,
  truncation warnings, const SRWLOCK bug; all 21 targets clean under /W4 /WX
- TIP loader: convention-agnostic key matching (leaf_0x4_sub0/leaf_0x4_0/
  plain), full 0x0..0x1F + 0x80000000-8 coverage, any [cpuid] config is
  authoritative (defaults wiped so no leaf leaks i9-10900K values); verified
  byte-exact against LUST donor capture
- WS-1 zero-rule: shared EvaluateLeaf policy (TIP -> zero, never native) used
  by exit handler and both WHP CPUID result lists; leaf-1 ECX[31]/SMX clear;
  max-leaf clamps; throttled zero-rule miss logging
- Close native-CPUID leak paths: AllocTracker VEH now uses the shared policy
  + RdtscHandler::ReadSpoofedTsc(); Partition no-handler fallback serves zeros
- CPUID exit-latency telemetry; CPUID->RDTSC pair compensation verified
- README: M0 + M1/WS-1 status, corrected syscall-bitmap claims, BYOD note

+++ src/engine/whp/Snapshot.cpp
@@ -227,7 +227,7 @@ bool Snapshot::RestoreInMemory(const std::vector<uint8_t>& snapshotData,
                 hOffset += msrHandler->Deserialize(hData + hOffset, hRemain - hOffset);
             }
             if (hRemain > hOffset && eptExecHook) {
-                eptExecHook->Deserialize(hData + hOffset, hRemain - hOffset);
+                eptExecHook->Deserialize(hData + hOffset, (uint32_t)(hRemain - hOffset));
             }
         }
     }
@@ -426,7 +426,7 @@ bool Snapshot::Restore(const std::vector<uint8_t>& snapshotData,
                 hOffset += msrHandler->Deserialize(hData + hOffset, hRemain - hOffset);
             }
             if (hRemain > hOffset && eptExecHook) {
-                eptExecHook->Deserialize(hData + hOffset, hRemain - hOffset);
+                eptExecHook->Deserialize(hData + hOffset, (uint32_t)(hRemain - hOffset));
             }
         }
     }
